+++ test/core/workflow/mocks/text2image-body.html
@@ -5,62 +5,6 @@
       <div data-valign="middle" class="text">
       <h3 id="free-ai-image-generator" class="heading-xxl">Free AI Image Generator.</h3>
       <p class="body-m">Bring your ideas to life with Generate Image, the commercially safe text to image generator in Adobe Express powered by Adobe Firefly.</p>
-      <div class="ex-unity-wrap">
-        <div class="unity-sprite-container"></div>
-        <div class="ex-unity-widget">
-          <div class="autocomplete" role="combobox">
-            <div class="inp-wrap">
-              <input id="promptInput" class="inp-field" type="text" placeholder="Describe your image or click surprise me..." aria-autocomplete="list" aria-haspopup="listbox" aria-controls="prompt-dropdown" aria-expanded="false" aria-owns="prompt-dropdown" aria-activedescendant="" data-event-bound="true">
-              <div class="act-wrap">
-                <a href="#" class="unity-act-btn surprise-btn" data-event-bound="true" daa-ll="Surprise me-1--Free AI Image Genera">
-                  <div class="btn-ico">
-                    <img loading="lazy" src="/cc-shared/assets/img/product-features-icons/firefly/surprise-icon.svg" alt="">
-                  </div>
-                  <div class="btn-txt">Surprise me</div>
-                </a>
-                <a href="#" class="unity-act-btn gen-btn" data-event-bound="true" daa-ll="Generate now-2--Free AI Image Genera">
-                  <div class="btn-ico">
-                    <img loading="lazy" src="/cc-shared/assets/img/product-features-icons/firefly/generate-icon.svg" alt="">
-                  </div>
-                  <div class="btn-txt">Generate now</div>
-                </a>
-                <a href="#" class="unity-act-btn refresh-btn" data-event-bound="true" daa-ll="Refresh Button">
-                  <div class="btn-ico">
-                    <img loading="lazy" src="/cc-shared/assets/img/product-features-icons/firefly/generate-icon.svg" alt="">
-                  </div>
-                  <div class="btn-txt">Refresh Button</div>
-                </a>
-              </div>
-            </div>
-            <ul id="prompt-dropdown" class="drop hidden" daa-lh="Marquee" role="listbox" aria-labelledby="promptInput" aria-hidden="true">
-              <li class="drop-title-con" aria-labelledby="prompt-suggestions">
-                <span class="drop-title" id="prompt-suggestions">Prompt Suggestions</span>
-                <button class="close-btn" daa-ll="drop-close" aria-label="Close dropdown" data-event-bound="true">
-                  <svg><use xlink:href="#unity-close-icon"></use></svg>
-                </button>
-              </li>
-              <li id="item-0" class="drop-item" role="option" tabindex="0" aria-label="A wooden coffee machine, high quality, high resolution" aria-description="Prompt Suggestions" daa-ll="drop-cur-prompt-0|A wooden coffee machine, high quality, high resolution" data-event-bound="true">
-                <svg><use xlink:href="#unity-prompt-icon"></use></svg>Test Suggestion
-              </li>
-              <li id="item-1" class="drop-item" role="option" tabindex="0" aria-label="Stones, dried flowers neutral background" aria-description="Prompt Suggestions" daa-ll="drop-cur-prompt-1|Stones, dried flowers neutral background" data-event-bound="true">
-                <svg><use xlink:href="#unity-prompt-icon"></use></svg> Stones, dried flowers neutral background
-              </li>
-              <li id="item-2" class="drop-item" role="option" tabindex="0" aria-label="Latte, coffee mug, table, window, rain, ultra realistic, ultra resolution" aria-description="Prompt Suggestions" daa-ll="drop-cur-prompt-2|Latte, coffee mug, table, window, rain, ultra realistic, ultra resolution" data-event-bound="true">
-                <svg><use xlink:href="#unity-prompt-icon"></use></svg> Latte, coffee mug, table, window, rain, ultra realistic, ultra resolution
-              </li>
-              <li class="drop-sep" role="separator"></li>
-              <li class="drop-footer">
-                <div id="tip-content" class="tip-con" role="note" aria-label="Tip Describe a subject, an action, a place, a mood, or a style.">
-                </div>
-                <div class="legal-con">
-                  <a href="https://www.stage.adobe.com/legal/licenses-terms/adobe-gen-ai-user-guidelines.html" class="legal-text" daa-ll="Adobe Generative AI -4--Free AI Image Genera">Adobe Generative AI Terms</a>
-                </div>
-              </li>
-            </ul>
-          </div>
-        </div>
-        <div class="sr-only" aria-live="polite" aria-atomic="true"></div>
-      </div>
       <p class="body-m"><a href="https://www.stage.adobe.com/legal/licenses-terms/adobe-gen-ai-user-guidelines.html" target="_blank" daa-ll="Adobe Generative AI -5--Free AI Image Genera">Adobe Generative AI Terms</a></p>
     </div>
     </div>
@@ -95,3 +39,4 @@
     </div>
   </div>
 </div>
+<div class="global-footer"><ul><li>abc</li><li>def</li></ul></div>
