--- test/core/workflow/workflow-ai/workflow.ai.test.js
@@ -1,19 +1,0 @@
-/* eslint-disable quote-props */
-/* eslint-disable quotes */
-import { readFile } from '@web/test-runner-commands';
-import { expect } from '@esm-bundle/chai';
-
-const { default: init } = await import('../../../../unitylibs/blocks/unity/unity.js');
-document.body.innerHTML = await readFile({ path: '../mocks/text2image-body.html' });
-
-describe('Unity Text-To-Image Block', () => {
-  before(async () => {
-    const unityElement = document.querySelector('.unity');
-    await init(unityElement);
-  });
-
-  it('Unity Text to image block should be loaded', async () => {
-    const unityWidget = document.querySelector('.unity');
-    expect(unityWidget).to.exist;
-  });
-});
